Additional test fixes
- Fix more test compilation warnings on some configs
- Prevent xt_intr test from using high-priority interrupts

Tests passing

+++ demos/cadence/sim/common/application_code/cadence_code/xt_alloca.c
@@ -84,7 +84,7 @@ static void putstr(const char *s)
             outbyte('\r');
             outbyte('\n');
         }
-        else if (iscntrl(c) && c != '\r') {
+        else if (iscntrl((int)c) && c != '\r') {
             outbyte('^');
             outbyte('@' + c);
         }
